card: refactor "header" into css grid
post metadata and the svg separator got refactored to be inside a single
grid container to, hopefully, make alignments easier and more robust

still not 100% happy (added small right and left padding to read time and tags),
but i think this was a necessary step in the right direction. Should be easier to
improve over this css layout. Maybe add the title inside this as well,
it would be cleaner but it also works just fine right now. Maybe if I
find an elegant way to fix all this padding issues.