add draft support
draft posts will be hidden by defautl and in the future I will be able
to easily add a toggle button that flips their visibility. For now there
is no way to show the posts but they are still directly accessible through
their url. There is a shitty draft of the logic for the visibility
flipping button in draft_toggle.astro