fix: short posts fill screen
really short posts would show the underlying green background from the
body since they wouldnt go to the bottom of the screen.
With this new approach this should be fixed. It had the side effect of
breaking the navbar positioning so it was also altered to account for
this. Not extremelly happy with all body tags being flex containers but
it hasnt had any problematic side effect so far (besides the navbar)