fix build: not-prose not detected at build time
the not-prose selector works fine in dev but at build time it seems to
behave differently. The working hypothesis is that since not-prose is a
target with no styling associated, when we try to make the apply it
breaks down since there is nothing to be applied.

To fix this, i rolledback to the use of !important instead of not-prose.
Not my ideal solution but its better than fiddling with the html
generated from markdown