fix tags page on mobile & mv flex from base layout
I wasnt happy with having every body be a flex element by default but i
was to tired to notice that I had already made a prop to pass classes
into the body (previously only used for background). So i refactored
everything to make the flex class to be passed only in the [id].astro
pages. I had to keep the w-full in the navbar because it would get
squished inside the blog posts.

This also fixed the annoying bug in the tags page where the accordeons
wouldnt shrin after a certain size and break everything. Since I was
messing with the tags page I also fixed the posts display on mobile, I
hadnt realised that the date would squish the posts and make it almost
unreadable. I added some breakpoints to smooth out things and make it
more breathable. Besides that, I learned that there exists a time html
tag and that I can use css to dinamically change the content of a div
with ::after and ::before, neet stuff.