add slides from slides.com
This was hell and I will write more about it on the post about building
the website, but everythong uses the same js and css files and where
extracted from the zip exports from slides.com. However, I had to edit
the path for the css and js files inside the html files because if they
stayed inside src astro wouldnt serve them. In public they are readilly
available but we also loose astro processing of those files. Hope there
was not much to optimize