slides: final final -- fix imgs & pathing
added a more robust way to deal with the images, tha allows me to leave
the html asis when exported by intercepting the img elements and
updating their sources to check inside public.

maybe this could be a way to move them from public and get the image
processing from astro that we are loosing?

also, now idea why the link to the new slides weas behaving weirdly so i
changed it to an absolute path inside the optuna post and it seemed to
fix things. Honestly, no idea about this one and Im to tired to care,
everything seems fixed with no possible new side effects caused by this
so im calling quits for the day